Moved the contract to Illuminate\Routing\Contracts
